source_code/project7-fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
haversine==2.9.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
openai==1.58.1
packaging==24.2
pandas==2.2.3
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
wheel==0.44.0
